CMakeLists - re-order QT_ANDROID logic
When trying to build as a shared lib for Android on OSX, the current
logic results in the wrong includes.

A prime example of this is the following block:
```
if (MSVC)
  target_include_directories(_opencpn INTERFACE ${CMAKE_SOURCE_DIR}/buildwin)
elseif (APPLE)
  target_include_directories(
    _opencpn INTERFACE ${CMAKE_SOURCE_DIR}/buildosx/include
  )
elseif (QT_ANDROID)
  target_include_directories(_opencpn INTERFACE ${CMAKE_SOURCE_DIR}/android)
  target_link_libraries(${PACKAGE_NAME} PRIVATE "${CMAKE_CURRENT_BINARY_DIR}/_deps/lunasvg-build/liblunasvg.a")
  target_link_libraries(${PACKAGE_NAME} PRIVATE "${CMAKE_CURRENT_BINARY_DIR}/_deps/lunasvg-build/plutovg/libplutovg.a")
endif ()
```

Which results in lunasvg not being built at all.

Currently QT_ANDROID is the only variant implicitly cross compiling, so
moving that logic to the "top" of the if/else chain should allow the
overrides to work without changing the behaviour on other platforms.